ZarrDatasets

ZarrDatasets.ZarrDataset — Type

ds = ZarrDataset(url::AbstractString,mode = "r";
                 _omitcode = 404,
                 maskingvalue = missing)
ZarrDataset(f::Function,url::AbstractString,mode = "r";
                 maskingvalue = missing)

Open the zarr dataset at the url or path url. Only the read-mode is currently supported. ds supports the API of the JuliaGeo/CommonDataModel.jl. The experimental _omitcode allows to work-around servers that return HTTP error different than 404 for missing chunks.

Example:

using ZarrDatasets
url = "https://s3.waw3-1.cloudferro.com/mdl-arco-time-035/arco/MEDSEA_MULTIYEAR_PHY_006_004/med-cmcc-ssh-rean-d_202012/timeChunked.zarr"
ds = ZarrDataset(url);
# see the metadata
display(ds)
# load the variable time
time = ds["time"][:]
# load the the attribute long_name for the variable zos
zos_long_name = ds["zos"].attrib["long_name"]
# load the global attribute
comment = ds.attrib["comment"]
# query the dimension of the variable zos
size(ds["zos"])
close(ds)

Example with a do-block:

using ZarrDatasets
url = "https://s3.waw3-1.cloudferro.com/mdl-arco-time-035/arco/MEDSEA_MULTIYEAR_PHY_006_004/med-cmcc-ssh-rean-d_202012/timeChunked.zarr"

zos1 = ZarrDataset(url) do ds
  ds["zos"][:,:,end,1]
end # implicit call to close(ds)

Differences between Zarr and NetCDF files

All metadata is stored in JSON files for Zarr with the following implications:

JSON does not distinguish between integers and real numbers. They are all considered as generic numbers. Whole numbers are loaded as Int64 and decimal numbers Float64. It is not possible to store the number 1.0 as a real number.

The order of keys in a JSON document is undefined. It is therefore not possible to have a consistent ordering of the attributes or variables.